Implement changes requested in PR review
