--- src/Microsoft.AspNet.OData.Shared/Formatter/ContentTypeMediaTypeMapping.cs
@@ -1,17 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation.  All rights reserved.
-// Licensed under the MIT License.  See License.txt in the project root for license information.
-
-namespace Microsoft.AspNet.OData.Formatter
-{
-    public partial class ContentTypeMediaTypeMapping
-    {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="ContentTypeMediaTypeMapping"/> class.
-        /// <param name="mediaType">The media type.</param>
-        /// </summary>
-        public ContentTypeMediaTypeMapping(string mediaType)
-            : base(mediaType)
-        {
-        }
-    }
-}
