+++ src/Microsoft.AspNet.OData.Shared/Microsoft.AspNet.OData.Shared.projitems
@@ -67,7 +67,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)Formatter\ODataRawValueMediaTypeMapping.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Formatter\QueryStringMediaTypeMapping.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Formatter\Serialization\ODataSerializerProvider.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Formatter\ContentTypeMediaTypeMapping.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)GetNextPageHelper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)IPerRouteContainer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)CompatibilityOptions.cs" />
@@ -258,7 +257,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)Routing\Conventions\SelectControllerResult.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Routing\Conventions\SingletonRoutingConvention.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Routing\DefaultODataPathHandler.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Query\PlainTextODataQueryOptionsParser.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\IODataQueryOptionsParser.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Routing\ODataParameterValue.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Routing\DefaultODataPathValidator.cs" />
