+++ src/Microsoft.AspNetCore.OData/Query/PlainTextODataQueryOptionsParser.cs
@@ -5,11 +5,33 @@ using System.IO;
 using System.Threading.Tasks;
 using Microsoft.OData;
 using Microsoft.AspNet.OData.Common;
+using Microsoft.AspNetCore.Http;
+using System.Net.Http.Headers;
+using System;
 
 namespace Microsoft.AspNet.OData.Query
 {
-    public partial class PlainTextODataQueryOptionsParser
+    /// <summary>
+    /// Reads and parses the content of a <see cref="T:System.IO.Stream" /> 
+    /// into a query options part of an OData URL. 
+    /// The query options are passed in the request body as plain text.
+    /// </summary>
+    /// <remarks>This class derives from a platform-specific class.</remarks>
+    public partial class PlainTextODataQueryOptionsParser : IODataQueryOptionsParser
     {
+        private static MediaTypeHeaderValue supportedMediaType = MediaTypeHeaderValue.Parse("text/plain");
+
+        /// <inheritdoc/>
+        public bool CanParse(HttpRequest request)
+        {
+            if (request == null)
+            {
+                throw Error.ArgumentNull("httpRequest");
+            }
+
+            return request.ContentType?.StartsWith(supportedMediaType.MediaType, StringComparison.Ordinal) == true ? true : false;
+        }
+
         /// <inheritdoc/>
         public async Task<string> ParseAsync(Stream requestStream)
         {
